Always return the public IP address of the droplet
Right now the type of IP is inconsistent. If no private_networking is
specified when creating the droplet, it will return a public IP.
Otherwise it will return a private IP (because private IPs are listed
first by digital_ocean's API, and we are just returning the first
element in the json response).